Use forked zuul for added support for reconfiguration.
zuul-launcher needs an extra patch to reload logging configuration
on SIGUP (which is what zuul-launcher reconfigure fires). Use a forked
version which supports it.

Zuul v2.5 is no longer maintained, so the risk of us diverging from
upstream in time it takes us to upgrade to v3 is not that high.