Put zuul-executor ssh key on nodepool-builder
nodepool-builder has to install zuul-executor public SSH key on the
images it's building, so ensure that this key is available.

A byproduct of this commit is that logserver no longer can be used by
"old" clients that are using legacy zuulv2 keys (but that only affects
current development environment and has no effect on production).