Remove unused gerrit_ssh_host_key
We've introduced a hash zuul_ssh_host_keys used to add ssh known host
keys for all zuul connections, but kept gerrit_ssh_host_key in case
it was used for something. However, during merge to production that
variable has been deleted, and so it's verified to be of no use.