DNM: Preliminary Ubuntu Package Mirror
DO NOT MERGE
This is an initial implementation of the Ubuntu Package Mirror server
based on copying the logserver. It configures apt-mirror and Apache2
but it does NOT use the puppetforge apt_mirror module.

It also isn't entirely finished. The vhost template still has some
logserver specific bits that I haven't determined whether they need
to be replaced or merely deleted.